% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_rank_regression.R
\name{cca_rrr_cv_folds}
\alias{cca_rrr_cv_folds}
\title{Cross-validation Fold Evaluation for CCA_rrr}
\usage{
cca_rrr_cv_folds(
  X,
  Y,
  Sx,
  Sy,
  kfolds = 5,
  lambda = 0.01,
  r = 2,
  standardize = TRUE,
  solver = "ADMM",
  rho = 1,
  LW_Sy = TRUE,
  niter = 10000,
  thresh = 1e-04
)
}
\value{
Average RMSE across folds.
}
\description{
Evaluates a single value of lambda using k-fold CV.
}
