% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{FPR}
\alias{FPR}
\title{False Positive Rate (TPR)
This is a function that compares the structure of two matrices A and B.
It outputs the number of entries where A  is not zero but Bis.
A and B need to have the same number of rows and columns}
\usage{
FPR(A, B, tol = 1e-04)
}
\arguments{
\item{A}{A matrix.}

\item{B}{A matrix.}

\item{tol}{tolerance for declaring the entries non zero.}
}
\description{
False Positive Rate (TPR)
This is a function that compares the structure of two matrices A and B.
It outputs the number of entries where A  is not zero but Bis.
A and B need to have the same number of rows and columns
}
