% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{get_edge_incidence}
\alias{get_edge_incidence}
\title{Return the edge incidence matrix of an igraph graph}
\usage{
get_edge_incidence(g, weight = 1)
}
\arguments{
\item{g}{igraph graph object.}

\item{weight}{edge weights.}
}
\value{
Edge incidence matrix of the graph g, with +weight for the source node and -weight for the target node.
}
\description{
Return the edge incidence matrix of an igraph graph
}
