% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{principal_angles}
\alias{principal_angles}
\title{Calculate principal angles between subspace spanned by the columns of a and the subspace spanned by the columns of b}
\usage{
principal_angles(a, b)
}
\arguments{
\item{a}{A matrix whose columns span a subspace.}

\item{b}{A matrix whose columns span a subspace.}
}
\description{
Calculate principal angles between subspace spanned by the columns of a and the subspace spanned by the columns of b
}
