% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sinTheta}
\alias{sinTheta}
\title{SinTheta distance between subspaces}
\usage{
sinTheta(U, V)
}
\arguments{
\item{U}{A matrix whose columns span a subspace.}

\item{V}{A matrix whose columns span a subspace.}
}
\value{
sinTheta distance between the two subspaces spanned by the matrices A and B, defined as ||UU^T - VV^T||_F / sqrt(2)
}
\description{
Calculate the distance spanned by the columns of A and the subspace spanned by the columns of B, defined as ||UU^T - VV^T||_F / sqrt(2)
}
