% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{TPR}
\alias{TPR}
\title{True Positive Rate (TPR)
This is a function that compares the structure of two matrices A and B.
It outputs the number of entries that A and B have in common that are different from zero. A and B need to have the same number of rows and columns}
\usage{
TPR(A, B, tol = 1e-04)
}
\arguments{
\item{a}{A matrix.}

\item{b}{A matrix.}
}
\description{
True Positive Rate (TPR)
This is a function that compares the structure of two matrices A and B.
It outputs the number of entries that A and B have in common that are different from zero. A and B need to have the same number of rows and columns
}
