% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_Waaijenborg.R
\name{Waaijenborg}
\alias{Waaijenborg}
\title{Function to perform Sparse CCA based on Waaijenborg et al. (2008)
REFERENCE Waaijenborg et al. (2008), "Quantifying the Association between Gene Expressions and DNA-Markers by Penalized Canonical Correlation Analysis" in Statistical Applications in Genetics and Molecular Biology, Volume 7, Issue 1, Article 3}
\usage{
Waaijenborg(
  X,
  Y,
  lambdaxseq = matrix(seq(from = 1, to = 0, by = -0.01), nrow = 1),
  lambdayseq = matrix(seq(from = 1, to = 0, by = -0.01), nrow = 1),
  rank,
  selection.criterion = 1,
  n.cv = 5,
  max.iter = 20,
  conv = 10^-3,
  standardize = TRUE
)
}
\arguments{
\item{X}{Matrix of predictors (n x p)}

\item{Y}{Matrix of responses (n x q)}

\item{lambdaxseq}{Vector of sparsity parameters for X (default is a sequence from 0 to 1 with step 0.1)}

\item{lambdayseq}{Vector of sparsity parameters for Y (default is a sequence from 0 to 1 with step 0.1)}

\item{rank}{Number of canonical components to extract}

\item{selection.criterion}{Criterion for selecting the optimal tuning parameter (1 for minimizing difference between test and training sample correlation, 2 for maximizing test sample correlation)}

\item{n.cv}{Number of cross-validation folds (default is 5)}
}
\value{
A list with elements:
\describe{
\item{U}{Canonical direction matrix for X (p x r)}
\item{V}{Canonical direction matrix for Y (q x r)}
\item{loss}{Mean squared error of prediction}
\item{cor}{Canonical covariances}
}
}
\description{
Function to perform Sparse CCA based on Waaijenborg et al. (2008)
REFERENCE Waaijenborg et al. (2008), "Quantifying the Association between Gene Expressions and DNA-Markers by Penalized Canonical Correlation Analysis" in Statistical Applications in Genetics and Molecular Biology, Volume 7, Issue 1, Article 3
}
