% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_rank_regression.R
\name{cca_rrr_cv}
\alias{cca_rrr_cv}
\title{Cross-validated Canonical Correlation Analysis via RRR}
\usage{
cca_rrr_cv(
  X,
  Y,
  r = 2,
  Kx = NULL,
  lambda_Kx = 0,
  param_lambda = 10^seq(-3, 1.5, length.out = 100),
  kfolds = 14,
  solver = "ADMM",
  parallelize = FALSE,
  LW_Sy = FALSE,
  do.scale = TRUE,
  rho = 1,
  niter = 10000,
  thresh = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{X}{Matrix of predictors.}

\item{Y}{Matrix of responses.}

\item{r}{Rank of the solution.}

\item{Kx}{Optional penalty matrix for X.}

\item{lambda_Kx}{Regularization strength on penalty matrix Kx.}

\item{param_lambda}{Sequence of lambda values for cross-validation.}

\item{kfolds}{Number of folds for cross-validation.}

\item{solver}{Solver type: "rrr", "CVX", or "ADMM".}

\item{parallelize}{Logical; should cross-validation be parallelized?}

\item{LW_Sy}{Whether to use Ledoit-Wolf shrinkage for Sy.}

\item{do.scale}{Logical; should X and Y be scaled.}

\item{rho}{ADMM parameter.}

\item{niter}{Maximum number of iterations for ADMM.}

\item{thresh}{Convergence threshold.}

\item{verbose}{Logical for verbose output.}

\item{Sx}{Optional X covariance matrix.}

\item{Sy}{Optional Y covariance matrix.}

\item{lambda}{Regularization parameter.}

\item{highdim}{Boolean for high-dimensional regime.}
}
\value{
A list with elements:
\describe{
\item{U}{Canonical direction matrix for X (p x r)}
\item{V}{Canonical direction matrix for Y (q x r)}
\item{lambda}{Optimal regularisation parameter lambda chosen by CV}
\item{rmse}{Mean squared error of prediction (as computed in the CV)}
\item{cor}{Canonical correlations}
}
}
\description{
Performs cross-validation to select optimal lambda, fits CCA_rrr.
Canonical Correlation Analysis via Reduced Rank Regression (RRR)
}
