% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecca.r
\name{ecca.cv}
\alias{ecca.cv}
\title{Sparse Canonical Correlation via Reduced-Rank Regression when both X and Y are high-dimensional, with Cross-Validation}
\usage{
ecca.cv(
  X,
  Y,
  lambdas = 0,
  groups = NULL,
  r = 2,
  rho = 1,
  B0 = NULL,
  nfold = 5,
  select = "lambda.min",
  eps = 1e-04,
  maxiter = 500,
  verbose = F,
  parallel = F
)
}
\arguments{
\item{X}{Predictor matrix (n x p)}

\item{Y}{Response matrix (n x q)}

\item{lambdas}{Choice of regularization parameter}

\item{groups}{List of index vectors defining groups of predictors}

\item{r}{Target rank}

\item{rho}{ADMM parameter}

\item{B0}{Initial value for the coefficient matrix (optional)}

\item{nfold}{Number of cross-validation folds}

\item{select}{Which lambda to select: "lambda.min" or "lambda.1se"}

\item{eps}{Convergence threshold for ADMM}

\item{maxiter}{Maximum number of ADMM iterations}

\item{verbose}{Print diagnostics}

\item{parallel}{Whether to run cross-validation in parallel}

\item{do.scale}{Whether to scale variables}
}
\value{
A list with elements:
\describe{
\item{U}{Canonical direction matrix for X (p x r)}
\item{V}{Canonical direction matrix for Y (q x r)}
\item{cor}{Canonical covariances}
\item{loss}{The prediction error 1/n * \| XU - YV\|^2}
}
}
\description{
Performs group-sparse reduced-rank regression for CCA using either ADMM or CVXR solvers.
}
