% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{sinTheta}
\alias{sinTheta}
\title{sinTheta spanned by the columns of a and the subspace spanned by the columns of b}
\usage{
sinTheta(U, V)
}
\arguments{
\item{a}{A matrix whose columns span a subspace.}

\item{b}{A matrix whose columns span a subspace.}
}
\description{
sinTheta spanned by the columns of a and the subspace spanned by the columns of b
}
