% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{subdistance}
\alias{subdistance}
\title{Calculate subdistance between subspace spanned by the columns of a and the subspace spanned by the columns of b}
\usage{
subdistance(A, B)
}
\arguments{
\item{a}{A matrix whose columns span a subspace.}

\item{b}{A matrix whose columns span a subspace.}
}
\description{
Calculate subdistance between subspace spanned by the columns of a and the subspace spanned by the columns of b
}
