% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{TNR}
\alias{TNR}
\title{True Negative Rate (TPR)
This is a function that compares the structure of two matrices A and B.
It outputs the number of entries where A and B are both 0.
A and B need to have the same number of rows and columns}
\usage{
TNR(A, B, tol = 1e-04)
}
\arguments{
\item{a}{A matrix.}

\item{b}{A matrix.}
}
\description{
True Negative Rate (TPR)
This is a function that compares the structure of two matrices A and B.
It outputs the number of entries where A and B are both 0.
A and B need to have the same number of rows and columns
}
